requirements.txt
fastapi
uvicorn
pymongo
requests
python-dotenv
bcrypt

langchain
langchain-community
langchain-ollama
langchain-huggingface
langgraph

ollama

duckduckgo-search
tavily-python

pypdf
sentence-transformers
faiss-cpu
pymupdf